% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{adjust_schedule}
\alias{adjust_schedule}
\title{Adjust program schedule dates}
\usage{
adjust_schedule(program_schedule = NA, allowed_days = c(1, 3, 5))
}
\arguments{
\item{program_schedule}{A program schedule created with the program_schedule function}

\item{allowed_days}{An integer vector specifying which days of the week are allowable.}
}
\value{
A schedule for a weightlifting program in terms of \code{cycle, day, date, exercise, variant, set, reps, percentage, training max, weight}
}
\description{
Modifies a program schedule to only include days specified by user
}
