% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{available_programs}
\alias{available_programs}
\title{List of available weightlifting programming templates}
\usage{
available_programs()
}
\value{
A character vector of available programs
}
\description{
List of available weightlifting programming templates
}
\examples{
"novice_linear_progression" \%in\% available_programs()
}
