% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv_data.R
\name{is_valid_weightlifting_log}
\alias{is_valid_weightlifting_log}
\title{Checks to see if a data frame is a valid weightlifting log. Looks for names.}
\usage{
is_valid_weightlifting_log(weightlifting.log = NULL)
}
\arguments{
\item{weightlifting.log}{A data frame containing at least the following elements: \code{program, date, exercise, variant, reps,  weight}}
}
\value{
Boolean
}
\description{
Checks to see if a data frame is a valid weightlifting log. Looks for names.
}
