% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv_data.R
\name{load_csv_data}
\alias{load_csv_data}
\title{Loads weightlifting logs in CSV format into a data frame}
\usage{
load_csv_data(files = NA, datadir = NA, header = TRUE)
}
\arguments{
\item{files}{A list of files containing weightlifting logs in CSV format. Not currently implemented.}

\item{datadir}{A directory containing weightlifting logs in CSV format. The expected format is \code{date, exercise, variant, set1weight, set1reps, ..., setNweight, setNreps}}

\item{header}{Whether the CSV file contains a header. Passed to read.csv.}
}
\value{
A data frame containing a weightlifting log with one set of an exercise per row. The program name listed in the data frame will correspond to the name of the CSV file from which the data was read.
}
\description{
Loads weightlifting logs in CSV format into a data frame
}
