% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{lower_531_1backoff_upper_863_pyramid}
\alias{lower_531_1backoff_upper_863_pyramid}
\title{Programming template for Lower 5-3-1 with 1 Backoff, Upper 8-6-3 Pyramid}
\usage{
lower_531_1backoff_upper_863_pyramid(...)
}
\arguments{
\item{...}{Some program functions take arguments for percentage schemes; the ellipsis lets other programming functions accept those arguments without tripping error checks.}
}
\value{
a list of template elements, including \code{name, duration, rep-max} on which percentages are based, \code{schedule}. Also has a \code{deload_duration, deload_schedule}, if desired.
}
\description{
Programming template for Lower 5-3-1 with 1 Backoff, Upper 8-6-3 Pyramid
}
