% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{madcow}
\alias{madcow}
\title{Programming template for Madcow 5x5}
\usage{
madcow(set_interval = 0.125, increment_percentage = 0.025, ...)
}
\arguments{
\item{set_interval}{the percentage to ramp between sets; lower values are harder}

\item{increment_percentage}{the increase in weight percentage on the heavy day of the cycle}

\item{...}{Some program functions take arguments for percentage schemes; the ellipsis lets other programming functions accept those arguments without tripping error checks.}
}
\value{
a list of template elements, including \code{name, duration, rep-max} on which percentages are based, \code{schedule}.
}
\description{
Programming template for Madcow 5x5
}
