% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_maxes.R
\name{one_rep_max}
\alias{one_rep_max}
\title{one-rep max calculator}
\usage{
one_rep_max(weight, reps, method = "epley", verbose = FALSE)
}
\arguments{
\item{weight}{Actual weight lifted}

\item{reps}{Actual number of repetitions performed}

\item{method}{The estimation technique to use; a list is available with \code{\link{rep_max_formulas}}}

\item{verbose}{Defaults to false. If true, will message which estimation technique is used.}
}
\value{
A one-rep maximum estimate
}
\description{
Provides one-rep max given a single weight and rep
}
