% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_maxes.R
\name{percentages}
\alias{percentages}
\title{percentages calculator}
\usage{
percentages(reps = c(2, 5, 8), weight = NA)
}
\arguments{
\item{reps}{A vector of reps}

\item{weight}{1RM on which to base percentages}
}
\value{
A table of 1RM percentage ranges
}
\description{
Provides a range of 1RM percentages for a given number of reps
}
