% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{program_schedule}
\alias{program_schedule}
\title{Program scheduling}
\usage{
program_schedule(
  program = NA,
  weightlifting.log = NA,
  smallest_plate = 2.5,
  increment = 0.025,
  cycles = 4,
  deload_every = 0,
  percentage = 0.9,
  prgm_start_date = Sys.Date() + 1,
  ...
)
}
\arguments{
\item{program}{The name of a supported program. Supported programs can be listed with \code{\link{available_programs}}}

\item{weightlifting.log}{A data frame containing at least the following elements: \code{program, date, exercise, variant, reps,  weight}}

\item{smallest_plate}{The smallest plate available to the trainee for the duration of the program. In English units, this is usually somewhere between 1.25 and 5 pounds.}

\item{increment}{The percentage to increment the weights over each cycle. 2.5\% is the default, which provides a 5-lb increase on upper body lifts and a 10-lb increase on lower body lifts for a typical male lifter.}

\item{cycles}{The number of cycles to include in the program.}

\item{deload_every}{If the program supports deloads, the number of cycles between each deload.}

\item{percentage}{Training max percentage. Defaults to 0.90.}

\item{prgm_start_date}{Date to start the program. Defaults to tomorrow.}

\item{...}{Variables to be passed to the \code{\link{training_max}} function, which establishes the baseline weight for the program template}
}
\value{
A schedule for a weightlifting program in terms of \code{cycle, day, exercise, variant, set, reps, percentage, training max, weight}
}
\description{
Creates a programming schedule from a program template given a user's historical information.
}
