% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_maxes.R
\name{rep_max_formulas}
\alias{rep_max_formulas}
\title{Available rep-max formulas}
\usage{
rep_max_formulas()
}
\value{
A character vector of available estimation techiques; formulas sourced from https://en.wikipedia.org/wiki/One-repetition_maximum
}
\description{
List of available estimation techniques for an arbitrary number of repetitions
}
\examples{
"wathan" \%in\% rep_max_formulas()

}
