% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_maxes.R
\name{training_max}
\alias{training_max}
\title{training max calculator}
\usage{
training_max(
  weightlifting.log = NA,
  program = NA,
  percentage = 0.9,
  method = "epley"
)
}
\arguments{
\item{weightlifting.log}{A data frame containing at least the following elements: \code{program, date, exercise, variant, reps,  weight}}

\item{program}{The name of a supported program. Supported programs can be listed with \code{\link{available_programs}}}

\item{percentage}{The percentage of estimated one-rep max at which to set the training max. Defaults to 90\%.}

\item{method}{The estimation technique to use; a list is available with \code{\link{rep_max_formulas}}}
}
\value{
A table of training maxes for each \code{exercise, equipment, variant} combination specified in the program template
}
\description{
An everyday training maximum commonly used to set percentages for weight programs
}
