% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_maxes.R
\name{n_rep_max}
\alias{n_rep_max}
\title{n-rep max calculator}
\usage{
n_rep_max(one_RM, reps, method = "epley", verbose = FALSE)
}
\arguments{
\item{one_RM}{A one-rep max, either actual or estimated}

\item{reps}{The number of repetitions for which an n_rep_max is desired}

\item{method}{The estimation technique to use; a list is available with \code{\link{rep_max_formulas}}}

\item{verbose}{Defaults to false. If true, will message which estimation technique is used.}
}
\value{
An n-rep maximum estimate
}
\description{
Provides n-rep max given a one-rep max and the desired rep max
}
