% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_maxes.R
\name{top_sets}
\alias{top_sets}
\title{top set calculator}
\usage{
top_sets(weightlifting.log = NULL, use.method = NA)
}
\arguments{
\item{weightlifting.log}{A data frame containing at least the following elements: \code{program, date, exercise, variant, reps,  weight}}

\item{use.method}{The estimation technique to use; a list is available with \code{\link{rep_max_formulas}}}
}
\value{
A table of the top sets in the weightlifting.log
}
\description{
Provides a one-rep maximum for each lift in a program
}
