% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{wendler_531_pyramid}
\alias{wendler_531_pyramid}
\title{Programming template for Wendler 5-3-1 Pyramid}
\usage{
wendler_531_pyramid()
}
\value{
a list of template elements, including \code{name, duration, rep-max} on which percentages are based, \code{schedule}. Also has a \code{deload_duration, deload_schedule}, if desired.
}
\description{
Programming template for Wendler 5-3-1 Pyramid
}
